+++ eventHandler/main.js
@@ -1,8 +1,7 @@
 import { fetchData } from "./view.js";
 import { historyEvent } from "./historyList.js";
-import { eventAddCard } from "./column.js";
+import { eventAddCard, update } from "./column.js";
 import { eventDelete } from "./deleteForm.js";
-import { update } from "./updateForm.js";
 
 historyEvent();
 fetchData();
